Функции выводит количество фильтров в верстке
