--- src/components/filter-data.js
@@ -1,26 +1,0 @@
-const filterData = [
-  {
-    title: `All`,
-    count: tasksMockData.length,
-  },
-  {
-    title: `Overdue`,
-  },
-  {
-    title: `Today`,
-  },
-  {
-    title: `Favorites`,
-  },
-  {
-    title: `Repeating`,
-  },
-  {
-    title: `Tags`,
-  },
-  {
-    title: `Archive`,
-  },
-];
-
-export {filterData};
